pyproject.toml
[project]
name = "meltwater-feeds-workers"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "httpx>=0.28.1",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
]
